requirements.txt
alabaster==0.7.12
altair==4.2.0
asttokens==2.1.0
attrs==22.1.0
awscli==1.27.7
Babel==2.11.0
backcall==0.2.0
black==22.10.0
blinker==1.5
botocore==1.29.7
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.0
colorama==0.4.4
commonmark==0.9.1
coverage==6.5.0
dask==2022.10.2
debugpy==1.6.3
decorator==5.1.1
docutils==0.16
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
Fiona==1.8.22
flake8==5.0.4
fsspec==2022.11.0
geopandas==0.12.1
gitdb==4.0.9
GitPython==3.1.29
google-api-core==2.10.2
google-auth==2.14.1
google-cloud-bigquery==3.3.6
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.50.0
grpcio-status==1.50.0
idna==3.4
imagesize==1.4.1
importlib-metadata==5.0.0
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.0
jupyter_client==7.4.5
jupyter_core==5.0.0
locket==1.0.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
munch==2.5.0
mypy-extensions==0.4.3
nbformat==5.7.0
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parso==0.8.3
partd==1.3.0
pathlib2==2.3.7.post1
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.3
plotly==5.11.0
prompt-toolkit==3.0.32
proto-plus==1.22.1
protobuf==4.21.9
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pydeck==0.8.0
pyflakes==2.5.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyproj==3.4.0
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
pyzmq==24.0.1
requests==2.28.1
rich==12.6.0
rsa==4.7.2
s3transfer==0.6.0
scikit-learn==1.1.3
scipy==1.9.3
semver==2.13.0
Shapely==1.8.5.post1
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.0
streamlit==1.14.0
tenacity==8.1.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
zipp==3.10.0
-e .